+++ NEWS.md
@@ -4,8 +4,10 @@
 
 Bug fixes:
  * dbus_attr_accessor and friends validate the signature ([#120][]).
+ * Declare the Introspectable interface in exported objects([#99][]).
 
 [#120]: https://github.com/mvidner/ruby-dbus/issues/120
+[#99]: https://github.com/mvidner/ruby-dbus/issues/99
 
 ## Ruby D-Bus 0.18.1 - 2022-07-13